# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AudienceRoom API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ai-characters": {
      "get": {
        "tags": [
          "ai-characters"
        ],
        "summary": "List Ai Characters",
        "operationId": "list_ai_characters_ai_characters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AiCharacterResponse"
                  },
                  "type": "array",
                  "title": "Response List Ai Characters Ai Characters Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ai-characters"
        ],
        "summary": "Create Ai Character",
        "operationId": "create_ai_character_ai_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiCharacterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiCharacterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ai-characters/{character_id}": {
      "get": {
        "tags": [
          "ai-characters"
        ],
        "summary": "Get Ai Character",
        "operationId": "get_ai_character_ai_characters__character_id__get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiCharacterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/session-messages": {
      "post": {
        "tags": [
          "session-messages"
        ],
        "summary": "Create Session Message",
        "operationId": "create_session_message_session_messages_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionMessageCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "session-messages"
        ],
        "summary": "List Session Messages",
        "operationId": "list_session_messages_session_messages_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionMessageResponse"
                  },
                  "title": "Response List Session Messages Session Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session-messages/{message_id}": {
      "get": {
        "tags": [
          "session-messages"
        ],
        "summary": "Get Session Message",
        "operationId": "get_session_message_session_messages__message_id__get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session-feedback": {
      "post": {
        "tags": [
          "session-feedback"
        ],
        "summary": "Create Session Feedback",
        "operationId": "create_session_feedback_session_feedback_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionFeedbackCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "session-feedback"
        ],
        "summary": "Get Session Feedback By Session",
        "operationId": "get_session_feedback_by_session_session_feedback_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session-feedback/{feedback_id}": {
      "get": {
        "tags": [
          "session-feedback"
        ],
        "summary": "Get Session Feedback",
        "operationId": "get_session_feedback_session_feedback__feedback_id__get",
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feedback Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/conversation/message": {
      "post": {
        "tags": [
          "conversation"
        ],
        "summary": "Send Message",
        "description": "Send a message to the AI and receive a response.\n\nThis endpoint processes the user's spoken message, generates an AI response\nusing the appropriate prompt for the session mode, and optionally returns\nsynthesized audio using VOICEVOX.",
        "operationId": "send_message_conversation_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiCharacterCreateRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "strictness": {
            "type": "string",
            "title": "Strictness"
          },
          "personality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personality"
          },
          "voice_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Style"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "role",
          "strictness"
        ],
        "title": "AiCharacterCreateRequest"
      },
      "AiCharacterResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "strictness": {
            "type": "string",
            "title": "Strictness"
          },
          "personality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personality"
          },
          "voice_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Style"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "role",
          "strictness",
          "personality",
          "voice_style",
          "description",
          "is_active"
        ],
        "title": "AiCharacterResponse"
      },
      "ConversationMessageRequest": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "generate_audio": {
            "type": "boolean",
            "title": "Generate Audio",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "ConversationMessageRequest"
      },
      "ConversationResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "audio_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Base64"
          },
          "speaker_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Id"
          },
          "participant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Id"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ConversationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionFeedbackCreateRequest": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "summary_title": {
            "type": "string",
            "title": "Summary Title"
          },
          "positive_points": {
            "title": "Positive Points"
          },
          "improvement_points": {
            "title": "Improvement Points"
          },
          "short_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Comment"
          },
          "closing_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Message"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "summary_title",
          "positive_points",
          "improvement_points"
        ],
        "title": "SessionFeedbackCreateRequest"
      },
...
      "SessionFeedbackResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "summary_title": {
            "type": "string",
            "title": "Summary Title"
          },
          "short_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Comment"
          },
          "positive_points": {
            "title": "Positive Points"
          },
          "improvement_points": {
            "title": "Improvement Points"
          },
          "closing_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "summary_title",
          "short_comment",
          "positive_points",
          "improvement_points",
          "closing_message",
          "created_at",
          "updated_at"
        ],
        "title": "SessionFeedbackResponse"
      },
...
      "SessionMessageCreateRequest": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "participant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Id"
          },
          "sequence_no": {
            "type": "integer",
            "title": "Sequence No"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "transcript_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Confidence"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "sequence_no",
          "content"
        ],
        "title": "SessionMessageCreateRequest"
      },
      "SessionMessageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "participant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Id"
          },
          "sequence_no": {
            "type": "integer",
            "title": "Sequence No"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "transcript_confidence": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Confidence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "participant_id",
          "sequence_no",
          "content",
          "transcript_confidence",
          "created_at"
        ],
        "title": "SessionMessageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}